Remove the need for legacy db volume
Using the old mechanism is not robust enough as the supervisor could be
running on a older OS still with the old `resin-supervisor` directory or
service name. The legacy volume also interferes with local testing on
some platforms.

The new mechanism uses the docker CLI and the existing supervisor image
to write to the database using JavaScript.

Change-type: patch