Write/read GET request cache on disk
This modifies the Get client in request to store a cache of the
response body when the response includes an ETag header. The cache includes
both the body and etag and is stored using the endpoint as the
filename. This prevents a restarting/crashing process to continuously
poll the API.

This also updates the core loop to use the cached state if any as the
first target. This allows the initial target to be applied even if there is not
network connection when the service start (e.g after a boot).

Change-type: patch