Break up main seek loop into separate actions
The original motivation for this change was to get out of tokio::select as soon as possible because it causes issues with automatic code formatting and inference.

But breaking it up this way can also help towards ultimately capturing the loop’s state machine in a type, which should make it more robust and easier to work on.

Change-type: patch