Add HTTP client
Wraps reqwest.

For the common case, works directly with (de)serializable types and JSON, skips much of the ceremony around making requests and bridges http::Uri with reqwest’s custom type which should ultimately result in less back and forth via Strings.

For the less common use case, it exposes a method to fully customize the request and gives access to the response.

Change-type: patch