Block API set_app_target until target is accepted
This involves the creation of a new ack_watch channel, that adds a
`send_and_wait` function that gets blocked until the receiver accepts
the request. The channel is configured in such a way that if the request
gets dropped (because it got replaced with a new request), then the
`send_and_wait` method returns an error.

The API `POST /v3/device/apps/{uuid}` endpoint will error with a 409
Conflict in case the state got replaced.

This should make it easier for users of the API to know when the
provided state got accepted by the worker or when it won't be processed.

Change-type: minor