Refactor Host type into OperatingSystem
The new OperatingSystem type parses a OS version string, e.g. `balenaOS 6.5.3+rev2` into the
corresponding parts.

This also removes the arch property, the architecture is a property of the device
type and is used as container metadata and for contracts in the legacy
supervisor. We'll re-evaluate adding it at that point. This commit also
removes the `device_type` property from the `Device` type for the same
reason.

Change-type: minor