Add a task to normalize service image info
The image uri read for the service may not fully match the image on the
target state as the digest may be missing in the engine image artifact
(due to deltas).

This just updates the in-memory service in that case to allow the
planner to reach the target.

Change-type: patch