Implement `store` module in helios-util
This module provides a simple implementation for a filesystem backed
store for serializable data. This store will be used to store data
coming from the backend that doesn't have a counterpart on the local
state of the system.

This commit modifies the `config` implementation in helios-util to use
the filesystem Store.

Change-type: patch